% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceformats.r
\docType{data}
\name{asset.action.types.format}
\alias{asset.action.types.format}
\title{Action Types Format}
\format{
An object of class \code{col_spec} of length 2.
}
\usage{
asset.action.types.format
}
\description{
Formatting specifications for meta.action.types.service.
}
\keyword{internal}
