% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceformats.r
\docType{data}
\name{asset.sondes.format}
\alias{asset.sondes.format}
\title{Meta Sondes Format}
\format{
An object of class \code{col_spec} of length 2.
}
\usage{
asset.sondes.format
}
\description{
Formatting specifications for meta.contacts.service.
}
\keyword{internal}
