% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceformats.r
\docType{data}
\name{asset.standardsolution.types.format}
\alias{asset.standardsolution.types.format}
\title{Standard Solution Types Format}
\format{
An object of class \code{col_spec} of length 2.
}
\usage{
asset.standardsolution.types.format
}
\description{
Formatting specifications for asset.standardsolution.types.service.
}
\keyword{internal}
