% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceformats.r
\docType{data}
\name{asset.verificationinstrument.types.format}
\alias{asset.verificationinstrument.types.format}
\title{Verification Instrument Types Format}
\format{
An object of class \code{col_spec} of length 2.
}
\usage{
asset.verificationinstrument.types.format
}
\description{
Formatting specifications for
asset.verificationinstrument.types.service.
}
\keyword{internal}
