% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.r
\docType{data}
\name{base.url}
\alias{base.url}
\title{Database URL}
\format{
An object of class \code{list} of length 2.
}
\usage{
base.url
}
\description{
URLs for WQP test and production databases.
}
\keyword{internal}
