% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{basic_get}
\alias{basic_get}
\title{Basic GET}
\usage{
basic_get(service.url)
}
\arguments{
\item{service.url}{The complete URL of the web service.}
}
\value{
The query results.
}
\description{
Basic WQP GET interface.
}
\keyword{internal}
