% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{basic_post}
\alias{basic_post}
\title{Basic POST}
\usage{
basic_post(service.url, body)
}
\arguments{
\item{service.url}{The complete URL of the web service.}

\item{body}{The POST field(s) to include in the request body.}
}
\value{
The server response.
}
\description{
Basic WQP POST interface.
}
\keyword{internal}
