% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.r
\name{check_action_consistency}
\alias{check_action_consistency}
\title{Action Consistency}
\usage{
check_action_consistency(current.record, prior.record)
}
\arguments{
\item{current.record}{One-row dataframe of "current" action.}

\item{prior.record}{One-row dataframe of "current" action.}
}
\value{
\code{TRUE} if the action is consistent with the prior record,
\code{FALSE} otherwise.
}
\description{
Helper function to check if visit action sequence makes sense.
}
\keyword{internal}
