% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.r
\name{check_actions}
\alias{check_actions}
\title{Check Action Records}
\usage{
check_actions(records, program, database)
}
\arguments{
\item{records}{A dataset of event IDs, sonde IDs, location IDs,
and actions. These records are checked against existing data.}
}
\value{
The records dataset, with additional columns
with suffix \code{".existing"}.
}
\description{
Check Action Records
}
\keyword{internal}
