% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.r
\name{check_events}
\alias{check_events}
\title{Check Event Records}
\usage{
check_events(records, program, database)
}
\arguments{
\item{records}{A dataset of arrival times, station IDs, and
contact IDs. These records are checked against existing data.}
}
\value{
The records dataset, with additional columns
with suffix \code{".existing"}.
}
\description{
Check Event Records
}
\keyword{internal}
