% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.r
\name{check_existing}
\alias{check_existing}
\title{Existence Checker}
\usage{
check_existing(
  input.data,
  exist.col = "exists",
  what = "records",
  response = c("warning", "error")
)
}
\arguments{
\item{input.data}{The input dataset.}

\item{exist.col}{The field identifying entries as new or existing.}

\item{what}{A label for messaging.}
}
\value{
A vector of timestamps.
}
\description{
Helper function for identifying that all entries are new.
}
\keyword{internal}
