% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.r
\name{check_fields}
\alias{check_fields}
\alias{check_optional_fields}
\title{Field Checker}
\usage{
check_fields(required, provided, num.required = length(required))

check_optional_fields(optional, provided)
}
\arguments{
\item{required}{Vector of required fields.}

\item{provided}{Vector of provided fields.}

\item{num.required}{Number of fields required to be present.}

\item{optional}{Vector of optional fields.}
}
\description{
Check for missing fields.
}
\keyword{internal}
