% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.r
\name{check_ids}
\alias{check_ids}
\title{Lookup Checker}
\usage{
check_ids(label, name.col, id.col, response = c("warning", "error"))
}
\arguments{
\item{label}{The type of ID to check. Used for messaging.}

\item{name.col}{A vector of names.}

\item{id.col}{A vector of IDs.}
}
\description{
Helper function for verifying that all entries have an associated ID.
}
\keyword{internal}
