% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.r
\name{check_insert}
\alias{check_insert}
\title{Check Insertion}
\usage{
check_insert(
  input.data,
  response.col = "result",
  what = "records",
  na.action = c("warning", "error")
)
}
\arguments{
\item{input.data}{The input dataset.}

\item{response.col}{The field containing the server response.}

\item{what}{A label for messaging.}

\item{na.action}{The output response.}
}
\value{
\code{TRUE} if \code{response.cols} contain server errors,
\code{FALSE} otherwise.
}
\description{
Check that record insertion was successful.
}
\keyword{internal}
