% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.r
\name{check_na}
\alias{check_na}
\title{Identifier Checker}
\usage{
check_na(
  input.data,
  id.cols,
  what = "records",
  na.action = c("warning", "error")
)
}
\arguments{
\item{input.data}{The input dataset.}

\item{id.cols}{A vector of identifier column names.}

\item{what}{A label for messaging.}

\item{na.action}{The output response.}
}
\value{
\code{TRUE} if \code{id.cols} contain \code{NA} values, \code{FALSE} otherwise.
}
\description{
Helper function for verifying that all identifier records
are not \code{NA}.
}
\keyword{internal}
