% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.r
\name{check_prior_action}
\alias{check_prior_action}
\title{Check Prior Actions}
\usage{
check_prior_action(
  records,
  actions = c("Installed", "Removed"),
  program,
  database
)
}
\arguments{
\item{records}{Action records, i.e., output of
\code{\link[=wqp_format_actions]{wqp_format_actions()}}.}

\item{actions}{The action types to check for consistency. Default is
to check only "Installed" and "Removed" records.}

\item{program}{The program name.}

\item{database}{The database name.}
}
\value{
A logical vector identifying each action record as
consistent (\code{TRUE}) or not (\code{FALSE}).
}
\description{
Check if the specified actions are consistent with prior actions.
}
\keyword{internal}
