% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{check_result_data}
\alias{check_result_data}
\title{Check Result Records}
\usage{
check_result_data(records, program, database)
}
\arguments{
\item{records}{A dataset of result ids, timestamps, values, and flag
codes. These records are checked against existing data.}

\item{program}{The program name.}

\item{database}{The database name.}
}
\value{
The records dataset, with additional columns
\code{"values.existing"} and \code{"qaqc_flag_id.existing"}.
}
\description{
Check Result Records
}
\keyword{internal}
