% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardsolutions.r
\name{check_standard_solution_types}
\alias{check_standard_solution_types}
\title{Check Standard Solution Types}
\usage{
check_standard_solution_types(records, program, database)
}
\arguments{
\item{records}{A dataset of standard solutions.
These records are checked against existing data.}
}
\value{
The records dataset, with additional columns
with suffix \code{".existing"}.
}
\description{
Check Standard Solution Types
}
\keyword{internal}
