% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardsolutions.r
\name{check_standard_solutions}
\alias{check_standard_solutions}
\title{Check Standard Solution Records}
\usage{
check_standard_solutions(records, program, database)
}
\arguments{
\item{records}{A dataset of event IDs, solution IDs, lot
numbers, and expiration dates. These records are checked
against existing data.}
}
\value{
The records dataset, with additional columns
with suffix \code{".existing"}.
}
\description{
Check Standard Solution Records
}
\keyword{internal}
