% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.r
\name{check_timestamps}
\alias{check_timestamps}
\title{Timestamp Checker}
\usage{
check_timestamps(input.data, time.col, na.action = c("warning", "error"))
}
\arguments{
\item{input.data}{The input dataset.}

\item{time.col}{The field containing timestamps.}

\item{na.action}{The output response.}
}
\value{
A vector of timestamps.
}
\description{
Check and parse timestamps.
}
\keyword{internal}
