% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verificationinstruments.r
\name{check_verification_instrument_types}
\alias{check_verification_instrument_types}
\title{Check Verification Instrument Types}
\usage{
check_verification_instrument_types(records, program, database)
}
\arguments{
\item{records}{A dataset of verification instruments.
These records are checked against existing data.}
}
\value{
The records dataset, with additional columns
with suffix \code{".existing"}.
}
\description{
Check Verification Instrument Types
}
\keyword{internal}
