% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verificationinstruments.r
\name{check_verification_instruments}
\alias{check_verification_instruments}
\title{Check Verification Instrument Records}
\usage{
check_verification_instruments(records, program, database)
}
\arguments{
\item{records}{A dataset of event IDs, instrument IDs, serial
numbers, and calibration due dates. These records are checked
against existing data.}
}
\value{
The records dataset, with additional columns
with suffix \code{".existing"}.
}
\description{
Check Verification Instrument Records
}
\keyword{internal}
