% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{detect_sequence}
\alias{detect_sequence}
\title{Detect Sequence}
\usage{
detect_sequence(x, pattern)
}
\arguments{
\item{x}{A vector of values.}

\item{pattern}{The pattern to search for.}
}
\value{
A vector of same length as \code{x}, with unique integer
values for each detected sequence and \code{NA} elsewhere.
}
\description{
Detect a specific pattern in a vector.
}
\keyword{internal}
