% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceformats.r
\docType{data}
\name{event.action.details.format}
\alias{event.action.details.format}
\alias{empty.actions}
\title{Event Action Details Format}
\format{
An object of class \code{col_spec} of length 2.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 0 rows and 10 columns.
}
\usage{
event.action.details.format

empty.actions
}
\description{
Formatting specifications for event.action.details.service.
}
\keyword{internal}
