% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceformats.r
\docType{data}
\name{event.data.format}
\alias{event.data.format}
\alias{empty.events}
\title{Event Data Format}
\format{
An object of class \code{col_spec} of length 2.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 0 rows and 15 columns.
}
\usage{
event.data.format

empty.events
}
\description{
Formatting specifications for event.data.service.
}
\keyword{internal}
