% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceformats.r
\docType{data}
\name{event.reasons.format}
\alias{event.reasons.format}
\title{Event Reason Format}
\format{
An object of class \code{col_spec} of length 2.
}
\usage{
event.reasons.format
}
\description{
Formatting specifications for event.reasons.service.
}
\keyword{internal}
