% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceformats.r
\docType{data}
\name{event.standardsolution.details.format}
\alias{event.standardsolution.details.format}
\alias{empty.standardsolutions}
\title{Event Standard Solution Details Format}
\format{
An object of class \code{col_spec} of length 2.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 0 rows and 5 columns.
}
\usage{
event.standardsolution.details.format

empty.standardsolutions
}
\description{
Formatting specifications for
event.verificationinstrument.details.service.
}
\keyword{internal}
