% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceformats.r
\docType{data}
\name{event.summaries.format}
\alias{event.summaries.format}
\title{Event Summary Format}
\format{
An object of class \code{col_spec} of length 2.
}
\usage{
event.summaries.format
}
\description{
Formatting specifications for event.summaries.service.
}
\keyword{internal}
