% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceformats.r
\docType{data}
\name{event.types.format}
\alias{event.types.format}
\title{Event Type Format}
\format{
An object of class \code{col_spec} of length 2.
}
\usage{
event.types.format
}
\description{
Formatting specifications for event.types.service.
}
\keyword{internal}
