% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceformats.r
\docType{data}
\name{event.verificationinstrument.details.format}
\alias{event.verificationinstrument.details.format}
\alias{empty.verificationinstruments}
\title{Event Verification Instrument Details Format}
\format{
An object of class \code{col_spec} of length 2.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 0 rows and 5 columns.
}
\usage{
event.verificationinstrument.details.format

empty.verificationinstruments
}
\description{
Formatting specifications for
event.verificationinstrument.details.service.
}
\keyword{internal}
