% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{extract_response}
\alias{extract_response}
\title{Extract Response}
\usage{
extract_response(result, stop.on.fail = FALSE)
}
\arguments{
\item{result}{The result object, i.e. output of \code{curl_fetch_memory}.}
}
\description{
Extract Response
}
\keyword{internal}
