% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.r
\name{frm_check}
\alias{frm_check}
\title{Check for existing form data in WQP}
\usage{
frm_check(form, program, database)
}
\arguments{
\item{form}{Validated form data, i.e. output of \code{frm_format()}.
Must contain list column \code{event} and optional list columns
\code{result}, \code{action}, \code{instrument}, and \code{solution}.}

\item{program}{The program name.}

\item{database}{The database name.}
}
\value{
A tibble of insert query responses.
}
\description{
Check if data matching the supplied form already exists in WQP.
}
