% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.r
\name{frm_format}
\alias{frm_format}
\title{Format Form Data for WQP}
\usage{
frm_format(
  form,
  na.omit = FALSE,
  na.action = c("warning", "error"),
  program,
  database
)
}
\arguments{
\item{form}{Form data, i.e. output of \code{frm_read()}.
Must contain list column \code{event} and optional list columns
\code{result}, \code{action}, \code{instrument}, and \code{solution}.}

\item{na.omit}{If \code{TRUE}, omit \code{NA} records under specific
circumstances before validating. See "Details" For more
information.}

\item{na.action}{How missing IDs are handled, i.e., data that cannot
be associated with an existing result ID will return either a
\code{warning} or an \code{error}.}

\item{program}{The program name.}

\item{database}{The database name.}
}
\value{
A tibble of formatted data, with additional columns
"validated" (logical) signaling if the form data was validated
and "log" containing information on validation failures.
}
\description{
Format all components of a form to the Water Quality Portal.
}
\details{
If \code{na.omit = TRUE}, records with \code{NA} values will be
removed under the following circumstances:
\itemize{
\item For \code{result} data, records where both \code{result_id} and \code{value}
are \code{NA} will be omitted.
\item For \code{action} data, records where \code{sonde_name} is \code{NA} will be
omitted.
\item For \code{instrument} data, records where \code{instrument_name} is \code{NA}
will be removed.
\item For \code{solution} data, records where \code{solution_name} is \code{NA}
will be removed.
Note that no \code{event} data records are removed.
}
}
