% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.r
\name{frm_insert}
\alias{frm_insert}
\title{Insert Form Data to WQP}
\usage{
frm_insert(form, overwrite = FALSE, program, database, token)
}
\arguments{
\item{form}{Validated form data, i.e. output of \code{frm_format()}.
Must contain list column \code{event} and optional list columns
\code{result}, \code{action}, \code{instrument}, and \code{solution}.}

\item{overwrite}{If \code{TRUE}, perform an update if the record already
exists.}

\item{program}{The program name.}

\item{database}{The database name.}

\item{token}{The token string.}
}
\value{
A tibble of insert query responses.
}
\description{
Insert all components of a form to the Water Quality Portal.
}
