% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.r
\name{frm_read}
\alias{frm_read}
\title{Read Event Sheet}
\usage{
frm_read(workbook, timezone = wqp_tz())
}
\arguments{
\item{workbook}{The Excel workbook (.xlsx) containing
the event data. See 'details' for the required format.}

\item{timezone}{The timezone assumed for the visit sheet.
Default is the standard WQP timezone, i.e. \code{"Etc/GMT+8"} (PST).}
}
\value{
A tibble of visit data.
}
\description{
Read event data from an electronic form.
}
\details{
The workbook is expected to contain one or more of the
following sheets:
\itemize{
\item \strong{EVENT}: this sheet is formatted to contain
columns that match the WQP event components:
\itemize{
\item \code{cdec_code}
\item \code{event_type_name}
\item \code{contact_name}
\item \code{arrival_time}
\item \code{restart_time}
\item \code{departure_time}
\item \code{reason_name}
\item \code{summary_name}
\item \code{comments}.
}
\item \strong{RESULT}: this sheet is formatted to contain
columns that match the WQP result components:
\itemize{
\item \code{cdec_code}
\item \code{interval_name}
\item \code{analyte_name}
\item \code{unit_name}
\item \code{rank_name}
\item \code{reading_type_name}
\item \code{time}
\item \code{value}
}
\item \strong{ACTION}: this sheet is formatted to contain columns
that match the WQP action components:
\itemize{
\item \code{arrival_time}
\item \code{cdec_code}
\item \code{contact_name}
\item \code{event_type_name}
\item \code{sonde_name}
\item \code{location_name}
\item \code{action_name}
}
\item \strong{INSTRUMENT}: this sheet is formatted to contain columns
that match the WQP verification instrument components:
\itemize{
\item \code{cdec_code}
\item \code{event_type_name}
\item \code{contact_name}
\item \code{arrival_time}
\item \code{instrument_name}
\item \code{serial_number}
\item \code{calibration_due_date}
}
\item \strong{SOLUTION}: this sheet is formatted to contain columns
that match the WQP standard solution components:
\itemize{
\item \code{cdec_code}
\item \code{event_type_name}
\item \code{contact_name}
\item \code{arrival_time}
\item \code{solution_name}
\item \code{lot_number}
\item \code{expiration_date}
}
}
}
