% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagner.r
\name{get_wagner_fields}
\alias{get_wagner_fields}
\title{Wagner Fields}
\usage{
get_wagner_fields(result.id, program, database)
}
\arguments{
\item{result.id}{The WQP result id.}

\item{program}{The program name.}

\item{database}{The database name.}
}
\value{
A dataframe of Wagner fields and corresponding result ids
associated with the specified time series.
}
\description{
Get the Wagner fields associated with a given time series result id.
}
\keyword{internal}
