% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.r
\name{getters}
\alias{getters}
\alias{get_contact_id}
\alias{get_station_id}
\alias{get_event_id}
\alias{get_event_type_id}
\alias{get_reason_id}
\alias{get_summary_id}
\alias{get_result_id}
\alias{get_reading_type_id}
\alias{get_sonde_id}
\alias{get_location_id}
\alias{get_action_id}
\alias{get_instrument_id}
\alias{get_solution_id}
\title{Getters}
\usage{
get_contact_id(new.data, program, database)

get_station_id(new.data, program, database)

get_event_id(new.data, program, database)

get_event_type_id(new.data, program, database)

get_reason_id(new.data, program, database)

get_summary_id(new.data, program, database)

get_result_id(new.data, program, database)

get_reading_type_id(new.data, program, database)

get_sonde_id(new.data, program, database)

get_location_id(new.data, program, database)

get_action_id(new.data, program, database)

get_instrument_id(new.data, program, database)

get_solution_id(new.data, program, database)
}
\arguments{
\item{new.data}{A dataframe.}

\item{program}{The program name.}

\item{database}{The database name.}
}
\value{
A vector of IDs.
}
\description{
internal functions for getting IDs based on other available fields.
}
\section{Functions}{
\itemize{
\item \code{get_contact_id}: Get Contact ID

\item \code{get_station_id}: Get Station ID

\item \code{get_event_id}: Get Event ID

\item \code{get_event_type_id}: Get Event Type ID

\item \code{get_reason_id}: Get Reason ID

\item \code{get_summary_id}: Get Summary ID

\item \code{get_result_id}: Get Result ID

\item \code{get_reading_type_id}: Get Reading Type ID

\item \code{get_sonde_id}: Get Sonde ID

\item \code{get_location_id}: Get Location ID

\item \code{get_action_id}: Get Action Type ID

\item \code{get_instrument_id}: Get Verification Instrument Type ID

\item \code{get_solution_id}: Get Standard Solution Type ID
}}

\keyword{internal}
