% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kor.r
\docType{data}
\name{kor.analyte.dictionary}
\alias{kor.analyte.dictionary}
\title{KOR Constituent Dictionary}
\format{
An object of class \code{character} of length 3.
}
\usage{
kor.analyte.dictionary
}
\description{
Field name equivalencies for KOR file and Suisun Marsh Analytes.
}
\keyword{internal}
