% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kor.r
\docType{data}
\name{kor.col.spec}
\alias{kor.col.spec}
\title{KOR Data Export Format}
\format{
An object of class \code{col_spec} of length 2.
}
\usage{
kor.col.spec
}
\description{
Formatting specifications for CSV files exported from KOR.
}
\keyword{internal}
