% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kor.r
\docType{data}
\name{kor.locale}
\alias{kor.locale}
\title{KOR Data Export Locale}
\format{
An object of class \code{locale} of length 7.
}
\usage{
kor.locale
}
\description{
Locale specifications for CSV files exported from KOR.
}
\keyword{internal}
