% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kor.r
\name{kor_format}
\alias{kor_format}
\title{KOR Data to WQP Records}
\usage{
kor_format(
  new.data,
  reading.type = "Time Series",
  interval = "15 min",
  na.action = c("warning", "error"),
  program,
  database
)
}
\arguments{
\item{new.data}{Timeseries data, i.e. output of \code{\link[=kor_read]{kor_read()}}.}

\item{reading.type}{The default reading type assumed for the KOR
data. Default is \code{"Time Series"}.}

\item{interval}{The default data interval assumed for the KOR
data. Default is \code{"15 min"}.}

\item{na.action}{How missing IDs are handled, i.e., data that cannot
be associated with an existing result ID will return either a
\code{warning} or an \code{error}.}

\item{program}{The program name.}

\item{database}{The database name.}
}
\value{
A tibble of records formatted for WQP.
}
\description{
Format result data for insertion to WQP, including the
identification of target result IDs.
}
\details{
This is mainly a wrapper for \code{\link{wqp_format_result_data}}
which sets some sensible defaults for result components.
}
