% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kor.r
\name{kor_read}
\alias{kor_read}
\title{Read KOR File}
\usage{
kor_read(file, cdec.regex = "^[A-Z]{3}", col.spec, locale.spec)
}
\arguments{
\item{file}{The CSV file exported from KOR.}

\item{cdec.regex}{Regular expression to extract the station CDEC
code from the filename.}

\item{col.spec}{The column format specifications for the file. Use
this to override the default expected format. See 'details' for
more information.}

\item{locale.spec}{The locale specifications for the file. Use
this to override the default expected locale. Default locale
assumes Latin character encoding and the standard WQP timezone,
i.e. \code{"Etc/GMT+8"} (PST).}
}
\value{
A tibble of time series data.
}
\description{
Read time series data from file exported from KOR.
}
\details{
The file is expected to the following columns:
\itemize{
\item Date (MM/DD/YYYY)
\item Time (HH:mm:ss)
\item Time (Fract. Sec)
\item Site Name
\item Columns for each analyte in format \verb{<analyte> <units>}.
}
}
