% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kor.r
\name{kor_trim}
\alias{kor_trim}
\title{Trim KOR Data}
\usage{
kor_trim(new.data, program, database)
}
\arguments{
\item{new.data}{The formatted KOR records, i.e. output of
\code{\link[=kor_format]{kor_format()}}.}

\item{program}{The program name.}

\item{database}{The database name.}
}
\value{
A subset of records from \code{new.data}.
}
\description{
Remove KOR data records that already exist in WQP. Useful for
extracting a subset of KOR records for backfilling data.
}
\details{
Records are selected based on presence/absence of
records in WQP.
}
