% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceformats.r
\docType{data}
\name{meta.contacts.format}
\alias{meta.contacts.format}
\title{Meta Contacts Format}
\format{
An object of class \code{col_spec} of length 2.
}
\usage{
meta.contacts.format
}
\description{
Formatting specifications for meta.contacts.service.
}
\keyword{internal}
