% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceformats.r
\docType{data}
\name{meta.locations.format}
\alias{meta.locations.format}
\title{Meta Locations Format}
\format{
An object of class \code{col_spec} of length 2.
}
\usage{
meta.locations.format
}
\description{
Formatting specifications for meta.locations.service.
}
\keyword{internal}
