% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceformats.r
\docType{data}
\name{meta.stations.format}
\alias{meta.stations.format}
\title{Meta Stations Format}
\format{
An object of class \code{col_spec} of length 2.
}
\usage{
meta.stations.format
}
\description{
Formatting specifications for meta.stations.service.
}
\keyword{internal}
