% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{multi_get}
\alias{multi_get}
\title{Multi GET}
\usage{
multi_get(service.url, stop.on.fail = TRUE)
}
\arguments{
\item{service.url}{The complete URL of the web service.}
}
\value{
The query results.
}
\description{
Asynchronous WQP GET interface.
}
\keyword{internal}
