% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{multi_post}
\alias{multi_post}
\title{Multi POST}
\usage{
multi_post(service.url, body, stop.on.fail = FALSE)
}
\arguments{
\item{service.url}{The complete URL of the web service.}

\item{body}{The POST field(s) to include in the request body.}
}
\value{
The server response.
}
\description{
Asynchronous WQP POST interface.
}
\keyword{internal}
