% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.r
\name{omit_na}
\alias{omit_na}
\title{Omit NA Values}
\usage{
omit_na(data, id.cols, choose = c("any", "all"))
}
\arguments{
\item{data}{The data to process.}

\item{id.cols}{Vector of columns to check for NA values. If missing,
all columns will be used.}

\item{choose}{Method for identifying rows to omit. If
\code{choose = "any"}, \code{NA} values in any of the columns specified by
\code{id.cols} will trigger an ommission. If \code{choose = "all"}, only rows
with \code{NA} values in all columns specified by \code{id.cols} will trigger
an omission.}
}
\value{
The data, with rows omitted.
}
\description{
Helper function for ommitting NA values. Provides greater
flexibility compared to \code{\link[stats:na.fail]{stats::na.omit()}}.
}
\keyword{internal}
