% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.r
\name{print_table}
\alias{print_table}
\title{Print Table}
\usage{
print_table(data, selection)
}
\arguments{
\item{data}{The data to concatenate for printing.}

\item{selection}{Logical vector of rows to subset from \code{data}.}
}
\value{
A string.
}
\description{
Helper function for printing error rows.
}
\keyword{internal}
