% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{process_datetime}
\alias{process_datetime}
\title{Process datetimes for WQP queries}
\usage{
process_datetime(x, default)
}
\arguments{
\item{x}{A vector that can be coerced to a datetime.}

\item{default}{The default value to use if \code{x} is missing.}
}
\value{
A vector of datetimes.
}
\description{
This function handles some timezone quirks.
}
\keyword{internal}
