% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.r
\docType{data}
\name{program.id}
\alias{program.id}
\title{Program ID}
\format{
An object of class \code{list} of length 3.
}
\usage{
program.id
}
\description{
Integer codes for WQP programs.
}
\keyword{internal}
