% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagner.r
\name{rating_table}
\alias{rating_table}
\title{Wagner Rating Table}
\usage{
rating_table(ratings)
}
\arguments{
\item{ratings}{list of Wagner ratings.}
}
\value{
A tibble.
}
\description{
Helper function to reformat Wagner specifications.
}
\keyword{internal}
