% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{read_results}
\alias{read_results}
\title{Read Results}
\usage{
read_results(content, result.cols, empty.data)
}
\arguments{
\item{content}{The result content, i.e. output of \code{basic_get()}}

\item{result.cols}{Result column type specification, i.e.,
\code{readr::cols()}, \code{readr::cols_only()}, etc.}
}
\value{
a tibble.
}
\description{
Read results into a dataframe.
}
\keyword{internal}
