% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{reformat_wqp_datetime}
\alias{reformat_wqp_datetime}
\title{Reformat WQP Datetime}
\usage{
reformat_wqp_datetime(x, space)
}
\arguments{
\item{x}{The datetime to reformat.}
}
\value{
The reformatted datetime.
}
\description{
Reformat datetime for WQP URLs.
}
\keyword{internal}
