% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceformats.r
\docType{data}
\name{report.pdm.format}
\alias{report.pdm.format}
\alias{empty.pdm.results}
\title{PDM Report Format}
\format{
An object of class \code{col_spec} of length 2.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 0 rows and 12 columns.
}
\usage{
report.pdm.format

empty.pdm.results
}
\description{
Formatting specifications for report.pdm.service.
}
\keyword{internal}
