% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceformats.r
\docType{data}
\name{result.data.format}
\alias{result.data.format}
\alias{empty.results}
\title{Result Format}
\format{
A \code{reader::cols()} specification or empty \code{tibble} object.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 0 rows and 5 columns.
}
\usage{
result.data.format

empty.results
}
\description{
Formatting specifications for web service Results/Result.
}
\keyword{internal}
