% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceformats.r
\docType{data}
\name{result.dates.format}
\alias{result.dates.format}
\alias{empty.result.dates}
\title{Result Dates Format}
\format{
An object of class \code{col_spec} of length 2.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 0 rows and 5 columns.
}
\usage{
result.dates.format

empty.result.dates
}
\description{
Formatting specifications for result.data.service.
}
\keyword{internal}
