% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceformats.r
\docType{data}
\name{result.details.format}
\alias{result.details.format}
\title{Result Details Format}
\format{
An object of class \code{col_spec} of length 2.
}
\usage{
result.details.format
}
\description{
Formatting specifications for result.details.service.
}
\keyword{internal}
