% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagner.r
\name{wag_drift_error}
\alias{wag_drift_error}
\title{Calculate Sensor Drift}
\usage{
wag_drift_error(
  event.id,
  interval.name = "Post-deployment check",
  reference.tag = "Calibration Standard",
  check.tag = "Post-deployment",
  rank.tag = "Standard",
  program,
  database,
  version
)
}
\arguments{
\item{event.id}{A vector of event IDs.}

\item{interval.name}{The interval name, used to filter the result
data based on the event type.}

\item{reference.tag}{A regex string to identify the calibration
standard reference value.}

\item{check.tag}{A regex string to identify the post-deployment
check reading.}

\item{rank.tag}{A regex label to identify the ranks to use for
calculating drift (i.e., only calculate drift for ranks that
contain the word "Standard").}

\item{program}{The program name.}

\item{database}{The database name.}

\item{version}{The version number.}
}
\description{
Calculate sensor drift from post-deployment check data.
}
