% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagner.r
\name{wag_event_sequence}
\alias{wag_event_sequence}
\title{Wagner Event Sequences}
\usage{
wag_event_sequence(
  reason.name = "Wagner QAQC",
  action.types = c("Calibrated", "Installed", "Removed", "Checked"),
  program,
  database
)
}
\arguments{
\item{reason.name}{A "reason" indentifier to filter the number of
events to search through.}

\item{action.types}{Ordered vector of action types, defines the
event sequence to search for.}

\item{program}{The program name.}

\item{database}{The database name.}
}
\value{
a \code{tibble} with columns corresponding to each action in
\code{action.types}. Each row is a sequence of events where the actions
occurred.
}
\description{
Group events into the Wagner sequence.
}
