% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagner.r
\name{wag_fouling_error}
\alias{wag_fouling_error}
\title{Calculate Fouling Error}
\usage{
wag_fouling_error(
  event.id,
  interval.name = "Visit",
  dirty.tag = "Dirty",
  clean.tag = "Clean",
  deployed.tag = "Sonde",
  verification.tag = "Verification",
  program,
  database,
  version
)
}
\arguments{
\item{event.id}{A vector of event IDs.}

\item{interval.name}{The interval name, used to filter the
event data.}

\item{dirty.tag}{regex label to identify the dirty sonde
data components.}

\item{clean.tag}{regex label to identify the clean sonde
data components.}

\item{deployed.tag}{regex label to identify the deployed sonde
data components.}

\item{verification.tag}{regex label to identify the verification
sonde data components.}

\item{program}{The program name.}

\item{database}{The database name.}

\item{version}{The version number.}
}
\value{
A tibble of fouling calculations.
}
\description{
Calculate fouling error from pre/post cleaning visit data.
}
\details{
This function is highly sensitive to the labeling
convention used to differentiate clean/dirty deployed/verification
sonde data. The \code{reading_type_name} of the result data must be in
the following format (\verb{---} signifies ignored text):

\verb{<--->. <dirty/clean tag> <deployed/verification tag> <--->}
}
