% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagner.r
\name{wag_rating_meta}
\alias{wag_rating_meta}
\title{Wagner Rating Meta Data}
\usage{
wag_rating_meta(rating.data, start = "Installed", end = "Removed")
}
\arguments{
\item{rating.data}{A tibble of rating data,
i.e. output of \code{\link[=wag_ratings]{wag_ratings()}}.}

\item{start}{The column name containing event IDs
corresponding to the equipment deployment. Default
is \code{"Installed"}.}

\item{end}{The column name containing event IDs
corresponding to the equipment removal. Default
is \code{"Removed"}.}
}
\value{
The tibble \code{rating.data}, with additional columns
\code{"start_time"}, \code{"end_time"}, and \code{"cdec_code"}.
}
\description{
Define the time interval and station that ratings apply to.
}
