% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagner.r
\name{wag_ratings}
\alias{wag_ratings}
\title{Wagner Rating}
\usage{
wag_ratings(error.data, rating.spec, default.rating = "Unreliable")
}
\arguments{
\item{error.data}{Total error data, i.e. output of
\code{\link[=wag_total_error]{wag_total_error()}}.}

\item{rating.spec}{YAML specification defining Wagner ratings.}

\item{default.rating}{Default rating to use when data is missing or
specifications are not defined.}
}
\value{
A tibble of Wagner ratings.
}
\description{
Assign Wagner ratings to based on total error.
}
