% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagner.r
\name{wag_total_error}
\alias{wag_total_error}
\title{Calculate Total Error}
\usage{
wag_total_error(sequence.data, fouling.data, drift.data)
}
\arguments{
\item{sequence.data}{Event sequence data,
e.g. output from \code{\link[=wag_event_sequence]{wag_event_sequence()}}.}

\item{fouling.data}{Fouling error data,
e.g. output from \code{\link[=wag_fouling_error]{wag_fouling_error()}}.}

\item{drift.data}{Calibration drift error data,
e.g. output from \code{\link[=wag_drift_error]{wag_drift_error()}}.}
}
\value{
a tibble of total drift calculations.
}
\description{
Calculate total error from combined fouling and
calibration drift data.
}
