% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagner.r
\docType{data}
\name{wagner.fields}
\alias{wagner.fields}
\title{Wagner Field Names}
\format{
An object of class \code{character} of length 8.
}
\usage{
wagner.fields
}
\description{
Vector of field names containing Wagner method results.
}
\keyword{internal}
