% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.r
\name{wqp-asset-listings}
\alias{wqp-asset-listings}
\alias{wqp_sondes}
\alias{wqp_action_details}
\alias{wqp_action_types}
\alias{wqp_verification_instrument_types}
\alias{wqp_verification_instrument_details}
\alias{wqp_standard_solution_types}
\alias{wqp_standard_solution_details}
\title{Asset Listings}
\usage{
wqp_sondes(program, database)

wqp_action_details(program, database)

wqp_action_types(program, database)

wqp_verification_instrument_types(program, database)

wqp_verification_instrument_details(program, database)

wqp_standard_solution_types(program, database)

wqp_standard_solution_details(program, database)
}
\arguments{
\item{program}{The program name.}

\item{database}{The database name.}
}
\description{
Listings for Asset Data.
}
\section{Functions}{
\itemize{
\item \code{wqp_sondes}: List sondes.

\item \code{wqp_action_details}: List actions.

\item \code{wqp_action_types}: List action types.

\item \code{wqp_verification_instrument_types}: List verification instrument types.

\item \code{wqp_verification_instrument_details}: List verification instruments uses.

\item \code{wqp_standard_solution_types}: List standard solution types.

\item \code{wqp_standard_solution_details}: List verification instruments uses.
}}

