% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.r
\docType{data}
\name{wqp-asset-services}
\alias{wqp-asset-services}
\alias{asset.sondes.service}
\alias{asset.verificationinstrument.types.service}
\alias{asset.standardsolution.types.service}
\alias{asset.action.types.insert.service}
\alias{asset.verificationinstrument.types.insert.service}
\alias{asset.standardsolution.types.insert.service}
\title{Asset Services}
\format{
a URL part.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
asset.sondes.service

asset.verificationinstrument.types.service

asset.standardsolution.types.service

asset.action.types.insert.service

asset.verificationinstrument.types.insert.service

asset.standardsolution.types.insert.service
}
\description{
WQP asset services.
}
\keyword{internal}
