% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.r
\name{wqp-databases}
\alias{wqp-databases}
\alias{wqp_use_database}
\alias{wqp_active_database}
\title{WQP Database Selection}
\usage{
wqp_use_database(database, validate = TRUE)

wqp_active_database()
}
\arguments{
\item{database}{The database name.}

\item{validate}{If \code{TRUE}, validate the database selection.}
}
\description{
Switch between test and production database access.
}
