% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.r
\name{wqp-event-listings}
\alias{wqp-event-listings}
\alias{wqp_event_details}
\alias{wqp_event_types}
\alias{wqp_event_reasons}
\alias{wqp_event_summaries}
\title{Event Listings}
\usage{
wqp_event_details(program, database)

wqp_event_types(program, database)

wqp_event_reasons(program, database)

wqp_event_summaries(program, database)
}
\arguments{
\item{program}{The program name.}

\item{database}{The database name.}
}
\description{
Listings for Event Data.
}
\section{Functions}{
\itemize{
\item \code{wqp_event_details}: List events.

\item \code{wqp_event_types}: List event types.

\item \code{wqp_event_reasons}: List event reasons.

\item \code{wqp_event_summaries}: List event summaries.
}}

