% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.r
\docType{data}
\name{wqp-event-services}
\alias{wqp-event-services}
\alias{event.types.service}
\alias{event.reasons.service}
\alias{event.summaries.service}
\alias{event.detail.service}
\alias{event.action.detail.service}
\alias{event.verificationinstrument.detail.service}
\alias{event.standardsolution.detail.service}
\alias{event.insert.service}
\alias{event.action.insert.service}
\alias{event.verificationinstrument.insert.service}
\alias{event.standardsolution.insert.service}
\title{Event Services}
\format{
a URL part.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
event.types.service

event.reasons.service

event.summaries.service

event.detail.service

event.action.detail.service

event.verificationinstrument.detail.service

event.standardsolution.detail.service

event.insert.service

event.action.insert.service

event.verificationinstrument.insert.service

event.standardsolution.insert.service
}
\description{
WQP event services.
}
\keyword{internal}
