% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.r
\docType{data}
\name{wqp-help-services}
\alias{wqp-help-services}
\alias{help.service}
\alias{token.service}
\alias{valid.token.service}
\title{Help Services}
\format{
a URL part.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
help.service

token.service

valid.token.service
}
\description{
WQP help services.
}
\keyword{internal}
