% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.r
\docType{data}
\name{wqp-meta-services}
\alias{wqp-meta-services}
\alias{meta.contacts.service}
\alias{meta.stations.service}
\alias{meta.locations.service}
\alias{meta.action.types.service}
\title{Meta Services}
\format{
a URL part.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
meta.contacts.service

meta.stations.service

meta.locations.service

meta.action.types.service
}
\description{
WQP meta services.
}
\keyword{internal}
