% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.r
\name{wqp-programs}
\alias{wqp-programs}
\alias{wqp_use_program}
\alias{wqp_active_program}
\title{WQP Program Selection}
\usage{
wqp_use_program(program, validate = TRUE)

wqp_active_program()
}
\arguments{
\item{program}{The program name.}

\item{validate}{If \code{TRUE}, validate the program selection.}
}
\description{
Switch between WQP programs.
}
