% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.r
\docType{data}
\name{wqp-report-services}
\alias{wqp-report-services}
\alias{report.pdm.service}
\title{Report Services}
\format{
a URL part.

An object of class \code{character} of length 1.
}
\usage{
report.pdm.service
}
\description{
WQP report services.
}
\keyword{internal}
