% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.r
\name{wqp-reports}
\alias{wqp-reports}
\alias{wqp_report_pdm}
\title{WQP Reports}
\usage{
wqp_report_pdm(
  station.id,
  year,
  month,
  qaqc.flag = "*",
  program,
  database,
  bind = FALSE
)
}
\arguments{
\item{station.id}{Integer ID of station to query.}

\item{year}{The report year.}

\item{month}{The report month.}

\item{qaqc.flag}{Return data with the specified flag only.
Default is \code{"*"} for all flags.}

\item{program}{The program name.}

\item{database}{The database name.}

\item{bind}{If \code{TRUE}, bind the results of multiple queries into a
single data frame.}
}
\description{
Functions for generating reports with WQP.
}
\section{Functions}{
\itemize{
\item \code{wqp_report_pdm}: Progressive Daily Mean report.
}}

