% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.r
\name{wqp-result-listings}
\alias{wqp-result-listings}
\alias{wqp_result_details}
\alias{wqp_result_constituents}
\alias{wqp_result_reading_types}
\alias{wqp_result_dates}
\title{Result Listings}
\usage{
wqp_result_details(program, database)

wqp_result_constituents(program, database)

wqp_result_reading_types(program, database)

wqp_result_dates(result.id, program, database)
}
\arguments{
\item{program}{The program name.}

\item{database}{The database name.}

\item{result.id}{The WQP result id.}
}
\description{
Listings for Result Data.
}
\section{Functions}{
\itemize{
\item \code{wqp_result_details}: List result sets.

\item \code{wqp_result_constituents}: List constituents.

\item \code{wqp_result_reading_types}: List reading, interval, and
aggregate types.

\item \code{wqp_result_dates}: List result start/end dates.
}}

