% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.r
\docType{data}
\name{wqp-result-services}
\alias{wqp-result-services}
\alias{result.data.service}
\alias{result.detail.service}
\alias{result.dates.service}
\alias{result.data.insert.service}
\title{Result Services}
\format{
a URL part.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
result.data.service

result.detail.service

result.dates.service

result.data.insert.service
}
\description{
WQP result services.
}
\keyword{internal}
