% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.r
\name{wqp-tokens}
\alias{wqp-tokens}
\alias{wqp_use_token}
\alias{wqp_request_token}
\alias{wqp_check_token}
\title{WQP Token Definition}
\usage{
wqp_use_token(token, database, validate = TRUE)

wqp_request_token(database, ...)

wqp_check_token(token, database)
}
\arguments{
\item{token}{The token string.}

\item{database}{The database name.}

\item{validate}{If \code{TRUE}, validate the token.}

\item{...}{Additional arguments to \code{browseURL}.}
}
\description{
Get or set WQP token for authentication. Use \code{wqp_request_token}
to open the token request web page in a browser
(authentication required).
}
