% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.r
\name{wqp-versions}
\alias{wqp-versions}
\alias{validate_version}
\title{WQP Version Validation}
\usage{
validate_version(version, program)
}
\arguments{
\item{version}{The version number.}

\item{program}{The program name.}
}
\description{
Validate WQP versions based on program.
}
\keyword{internal}
