% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.r
\name{wqp_check_connection}
\alias{wqp_check_connection}
\title{Check Connection to WQP}
\usage{
wqp_check_connection(database)
}
\arguments{
\item{database}{The database name.}
}
\value{
\code{TRUE} if database can be reached, \code{FALSE} otherwise.
}
\description{
Check if WQP database can be accessed.
}
