% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.r
\name{wqp_events}
\alias{wqp_events}
\title{Events Query}
\usage{
wqp_events(event.id, program, database, bind = FALSE)
}
\arguments{
\item{event.id}{The WQP event id.}

\item{program}{The program name.}

\item{database}{The database name.}

\item{bind}{If \code{TRUE}, bind the results of multiple queries into a
single data frame.}
}
\value{
The query results.
}
\description{
Get data for the specified result id.
}
