% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{wqp_format_result_data}
\alias{wqp_format_result_data}
\title{New Results to WQP Records}
\usage{
wqp_format_result_data(
  new.data,
  na.action = c("warning", "error"),
  program,
  database
)
}
\arguments{
\item{new.data}{A dataframe of processed data. Must contain
data columns \code{"time"} and \code{"value"}, and additional identification
columns such as \code{"reading_type_name"}, \code{"analyte_name"},
\code{"cdec_code"}, etc.}

\item{na.action}{How missing IDs are handled, i.e., data that cannot
be associated with an existing result ID will return either a
\code{warning} or an \code{error}.}

\item{program}{The program name.}

\item{database}{The database name.}
}
\value{
A dataframe of records formatted for insertion into WQP using
\code{wqp_insert_result_data}
}
\description{
Format result data for insertion to WQP, including the
identification of target result IDs.
}
\seealso{
wqpr::list_all() wqpr::wqp_insert_result_data()
}
