% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verificationinstruments.r
\name{wqp_format_verification_instruments}
\alias{wqp_format_verification_instruments}
\title{New Verification Instrument to WQP Records}
\usage{
wqp_format_verification_instruments(
  new.data,
  na.action = c("warning", "error"),
  match.event = FALSE,
  program,
  database
)
}
\arguments{
\item{new.data}{A dataframe of processed data. Must contain
data column \strong{instrument_name} to identify the
standard solution type. If column \emph{event_id} is not provided
and \code{match.event} is \code{TRUE},
additional columns \strong{arrival_time}, \strong{cdec_code},
\strong{contact_name}, and \strong{event_type_name} must be provided to
identify the event ID.}

\item{na.action}{How missing IDs are handled, i.e., data that cannot
be associated with an existing result ID will return either a
\code{warning} or an \code{error}.}

\item{match.event}{If \code{TRUE}, search for an existing event to
associate with \code{new.data}.}

\item{program}{The program name.}

\item{database}{The database name.}
}
\value{
A dataframe of records formatted for insertion into WQP using
\code{wqp_insert_verification_instruments}
}
\description{
Format verification instruments for insertion to WQP, including the
identification of target event IDs.
}
\seealso{
wqpr::wqp_verification_instrument_types()
wqpr::wqp_verification_instrument_details()
}
