% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.r
\name{wqp_help}
\alias{wqp_help}
\title{WQP Online Help}
\usage{
wqp_help(database, ...)
}
\arguments{
\item{database}{The database name.}

\item{...}{Additional arguments to \code{browseURL}.}
}
\description{
Open the WQP online help page in a browser window.
}
