% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.r
\name{wqp_insert_actions}
\alias{wqp_insert_actions}
\title{Insert Action Records}
\usage{
wqp_insert_actions(records, overwrite = FALSE, program, database, token)
}
\arguments{
\item{records}{A dataset with 4 required columns, in order:
\strong{event_id}, \strong{sonde_id}, \strong{location_id}, \strong{action_id}.}

\item{overwrite}{included for consistency with other functions.
Due to database structure limitations, actions cannot be updated.
Setting \code{overwrite = TRUE} will result in an error.}

\item{program}{The program name.}

\item{database}{The database name.}

\item{token}{The token string.}
}
\value{
A data frame with the server response to each insert attempt.
}
\description{
Insert action records into WQP.
}
