% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.r
\name{wqp_insert_events}
\alias{wqp_insert_events}
\title{Insert Event Records}
\usage{
wqp_insert_events(records, overwrite = FALSE, program, database, token)
}
\arguments{
\item{records}{A dataset with 4 required columns and
6 optional columns, in order: \strong{arrival_time},
\strong{station_id}, \strong{contact_id}, \strong{event_type_id}, \emph{summary_id},
\emph{reason_id}, \emph{departure_time}, \emph{watch_time}, \emph{restart_time},
\emph{comments}.}

\item{overwrite}{If \code{TRUE}, perform an update if the record already
exists.}

\item{program}{The program name.}

\item{database}{The database name.}

\item{token}{The token string.}
}
\value{
A data frame with the server response to each insert attempt.
}
\description{
Insert event records into WQP.
}
