% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{wqp_insert_result_data}
\alias{wqp_insert_result_data}
\title{Insert Result Data Records}
\usage{
wqp_insert_result_data(records, overwrite = FALSE, program, database, token)
}
\arguments{
\item{records}{A dataset with 5 required columns, in order:
\strong{result_id}, \strong{time}, \strong{value}, \strong{qaqc_flag_id}, and
\strong{version}. \code{NA} flag entries will automatically be assigned
code \code{"U"}.}

\item{overwrite}{If \code{TRUE}, perform an update if the record already
exists.}

\item{program}{The program name.}

\item{database}{The database name.}

\item{token}{The token string.}
}
\value{
A data frame with the server response to each insert attempt.
}
\description{
Insert Result records into WQP.
}
