% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardsolutions.r
\name{wqp_insert_standard_solution_types}
\alias{wqp_insert_standard_solution_types}
\title{Insert Standard Solution Types}
\usage{
wqp_insert_standard_solution_types(
  records,
  overwrite = FALSE,
  program,
  database,
  token
)
}
\arguments{
\item{records}{A dataset with 1 required column and 1 optional
column, in order: \strong{solution}, \emph{notes}.}

\item{overwrite}{If \code{TRUE}, perform an update if the record already
exists.}

\item{program}{The program name.}

\item{database}{The database name.}

\item{token}{The token string.}
}
\value{
A data frame with the server response to each insert attempt.
}
\description{
Insert Standard Solution types into WQP.
}
