% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardsolutions.r
\name{wqp_insert_standard_solutions}
\alias{wqp_insert_standard_solutions}
\title{Insert Standard Solution Records}
\usage{
wqp_insert_standard_solutions(
  records,
  overwrite = FALSE,
  program,
  database,
  token
)
}
\arguments{
\item{records}{A dataset with 2 required columns and 2 optional
columns, in order:
\strong{event_id}, \strong{solution_id}, \emph{lot_number}, \emph{expiration_date}.}

\item{overwrite}{If \code{TRUE}, perform an update if the record already
exists.}

\item{program}{The program name.}

\item{database}{The database name.}

\item{token}{The token string.}
}
\value{
A data frame with the server response to each insert attempt.
}
\description{
Insert standard solution records into WQP.
}
