% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verificationinstruments.r
\name{wqp_insert_verification_instruments}
\alias{wqp_insert_verification_instruments}
\title{Insert Verification Instrument Records}
\usage{
wqp_insert_verification_instruments(
  records,
  overwrite = FALSE,
  program,
  database,
  token
)
}
\arguments{
\item{records}{A dataset with 2 required columns and 2 optional
columns, in order:
\strong{event_id}, \strong{instrument_id}, \emph{serial_number},
\emph{calibration_due_date}.}

\item{overwrite}{If \code{TRUE}, perform an update if the record already
exists.}

\item{program}{The program name.}

\item{database}{The database name.}

\item{token}{The token string.}
}
\value{
A data frame with the server response to each insert attempt.
}
\description{
Insert verification instrument records into WQP.
}
