% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{wqp_result_data}
\alias{wqp_result_data}
\title{Result Data Query}
\usage{
wqp_result_data(
  result.id,
  start.date,
  end.date,
  version,
  program,
  database,
  bind = FALSE
)
}
\arguments{
\item{result.id}{The WQP result id.}

\item{start.date}{The query start date. If missing, the earliest
available date is used.}

\item{end.date}{The query end date. If missing, the latest
available date is used.}

\item{version}{The version number.}

\item{program}{The program name.}

\item{database}{The database name.}

\item{bind}{If \code{TRUE}, bind the results of multiple queries into a
single data frame.}
}
\value{
The query results.
}
\description{
Get data for the specified result id.
}
