% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardsolutions.r
\name{wqp_standard_solutions}
\alias{wqp_standard_solutions}
\title{Standard Solutions Query}
\usage{
wqp_standard_solutions(event.id, program, database, bind = FALSE)
}
\arguments{
\item{event.id}{The WQP event id.}

\item{program}{The program name.}

\item{database}{The database name.}

\item{bind}{If \code{TRUE}, bind the results of multiple queries into a
single data frame.}
}
\value{
The query results.
}
\description{
Get data on standard solutions associated with the specified
event id.
}
