% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{wqp_update_result_flags}
\alias{wqp_update_result_flags}
\title{Update Flags}
\usage{
wqp_update_result_flags(records, overwrite = c("U"), program, database, token)
}
\arguments{
\item{records}{A dataset of result ids, timestamps, values, flag
codes, and versions. The data values are used to check for changes
to the data. No changes to the data are pushed to the database.}

\item{overwrite}{Which flag codes are allowed to be overwritten.
By default, only \code{"U"} (unchecked) flags can be overwritten.}

\item{program}{The program name.}

\item{database}{The database name.}

\item{token}{The token string.}
}
\value{
A data frame with the server response to each update attempt.
}
\description{
Update QAQC flags in WQP. This function does not modify the actual
data values.
}
