% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verificationinstruments.r
\name{wqp_verification_instruments}
\alias{wqp_verification_instruments}
\title{Verification Instruments Query}
\usage{
wqp_verification_instruments(event.id, program, database, bind = FALSE)
}
\arguments{
\item{event.id}{The WQP event id.}

\item{program}{The program name.}

\item{database}{The database name.}

\item{bind}{If \code{TRUE}, bind the results of multiple queries into a
single data frame.}
}
\value{
The query results.
}
\description{
Get data on verification instruments associated with the specified
event id.
}
