% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{package}
\name{wqpr}
\alias{wqpr}
\alias{wqpr-package}
\title{Interface to WQP}
\description{
This packages provides an interface to the California Department of
Water Resources (DWR) Water Quality Portal (WQP) database via the
Vista Web Services REST API. For DWR internal use only.
}
\section{Package options}{


wqpr uses the following \code{\link[=options]{options()}} to configure behavior:

\itemize{
\item \code{wqpr.database}: The default database to use on start
("test" or "production").
\item \code{wqpr.program}: The default program to use on start
("MARSH", "EMP", or "DAYFLOW").
\item \code{wqpr.token}: A named list of tokens to use for
authentication to each database ("test" and/or "production").
}
}

\author{
\strong{Maintainer}: Michael Koohafkan \email{Michael.Koohafkan@water.ca.gov}

}
