% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceformats.r
\docType{data}
\name{wqpr.tz}
\alias{wqpr.tz}
\alias{wqpr.locale}
\title{wqpr timezone}
\format{
An object of class \code{character} of length 1.

An object of class \code{locale} of length 7.
}
\usage{
wqpr.tz

wqpr.locale
}
\description{
The timezone used by WQP. Timestamps are in Pacific Standard
Time (PST) and daylight savings is ignored.
}
\keyword{internal}
