% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{wrap_quotes}
\alias{wrap_quotes}
\title{Wrap In Quotes}
\usage{
wrap_quotes(x)
}
\arguments{
\item{x}{Value to wrap in double quotes.}
}
\value{
The value \code{x}, converted to character and wrapped in
double quotes.
}
\description{
Helper function for wrapping a value in double quotes.
}
\keyword{internal}
